backend/requirements.txt
django>=5.0.0
channels[daphne]
djangorestframework   
djoser   
djangorestframework-simplejwt
# redis-channels
python-dotenv
psycopg2
